Tweak automation tracing
